requirements (1).txt
python-binance==1.0.19
websockets==12.0
aiohttp==3.9.5
python-dotenv==1.0.1
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
loguru==0.7.2
requests==2.32.3
